src/test/java/org/cvg/junit5_app/ejemplo/models/BancoTest.java
package org.cvg.junit5_app.ejemplo.models;

import org.junit.jupiter.api.Test;

import java.math.BigDecimal;

import static org.junit.jupiter.api.Assertions.*;

class BancoTest {

    @Test
    void transferirDineroCuentas() {
        Cuenta cuenta = new Cuenta("Jhon Doe", new BigDecimal("2500"));
        Cuenta cuenta2 = new Cuenta("Andres", new BigDecimal("1500.8989"));

        Banco banco = new Banco();
        banco.setNombre("BBVA");
        banco.transferir(cuenta2, cuenta, new BigDecimal(500));
        assertEquals("1000.8989", cuenta2.getSaldo().toPlainString());
        assertEquals("3000", cuenta.getSaldo().toPlainString());
    }

    @Test
    void testRelacionBancoCuentas() {
        Cuenta cuenta = new Cuenta("Jhon Doe", new BigDecimal("2500"));
        Cuenta cuenta2 = new Cuenta("Andres", new BigDecimal("1500.8989"));

        Banco banco = new Banco();
        banco.addCuenta(cuenta);
        banco.addCuenta(cuenta2);
        banco.setNombre("BBVA");

        banco.transferir(cuenta2, cuenta, new BigDecimal(500));
        assertEquals("1000.8989", cuenta2.getSaldo().toPlainString());
        assertEquals("3000", cuenta.getSaldo().toPlainString());

        assertEquals(2, banco.getCuentas().size());
        assertEquals("BBVA", cuenta.getBanco().getNombre());

        assertEquals("Andres", banco.getCuentas().stream()
                .filter( x -> x.getPersona().equals("Andres") )
                .findFirst()
                .get().getPersona());
        assertTrue(banco.getCuentas().stream()
                .anyMatch( x -> x.getPersona().equals("Andres") ));
    }
}